+++ app/src/main/java/so/codeweaver/muzei/ponies/derpi/DerpibooruImage.kt
@@ -1,10 +1,13 @@
-package so.codeweaver.muzei.ponies
+package so.codeweaver.muzei.ponies.derpi
 
 import android.net.Uri
+import androidx.room.Entity
+import androidx.room.PrimaryKey
 import com.google.android.apps.muzei.api.provider.Artwork
-import java.util.*
 
+@Entity
 data class DerpibooruImage(
+        @PrimaryKey
         val id: String,
         val image: String,
         val uploader: String,
@@ -26,28 +29,5 @@ data class DerpibooruImage(
             webUri = Uri.parse(siteUri)
             persistentUri = Uri.parse(fullImageUri)
         }
-    }
-}
-
-data class DerpibooruResult(
-        val total: Int,
-        val search: Array<DerpibooruImage>
-) {
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (javaClass != other?.javaClass) return false
-
-        other as DerpibooruResult
-
-        if (total != other.total) return false
-        if (!Arrays.equals(search, other.search)) return false
-
-        return true
-    }
-
-    override fun hashCode(): Int {
-        var result = total
-        result = 31 * result + Arrays.hashCode(search)
-        return result
     }
 }
